Program runs, but not all features work
The Program Should run now, I dont know why or how it got corrupted but it did.

Nevertheless, the program ATM does have faults and if you are pulling THIS COMMIT, you will have to insert the appropriate sprites into the appropriate locations.

Please Do check the README file in Assets/Editor/ReadME

Remember, at the moment the program will run, but the proper climate sprites and their temperatures wont be accurate, the river and coastal sprites dont work and you have to fix them. It should be an ez fix